+++ src/compositions/header/header.stories.tsx
@@ -27,12 +27,15 @@ export const Default: Story = {
       <>
         <AvatarButton
           onClick={fn()}
-          avatarProps={{
-            src: "/avatars/rory.jpg",
-            alt: "User avatar",
-            fallbackText: "RM",
-            size: "s",
-          }}
+          fallbackText="RM"
+          size="s"
+          imageElement={
+            <img
+              src="/avatars/rory.jpg"
+              alt="User avatar"
+              className="object-cover w-full h-full"
+            />
+          }
         />
         <IconButton
           label="Settings"